fix(warning): disable narrowing conversion through uniform initializer and add static_cast to function caller, fixing the function definition

+++ lib/include/application.h
@@ -137,5 +137,5 @@ private:
     void createTextureSampler();
     
     void createUniformBuffers();
-    void updateUniformBufferCamera(uint32_t currentImage, float deltaTime, int width, int height);
+    void updateUniformBufferCamera(uint32_t currentImage, float deltaTime, float width, float height);
 };
